Refactor processor into 2 generators
